new window added, new function to add a branch

+++ PHYS495_CircuitAnalyzer/src/circuit_logic/CircuitComponent.java
@@ -15,7 +15,7 @@ public abstract class CircuitComponent {
 		this.type = type;
 	}
 	
-	public double getValue() {
+	public float getValue() {
 		return value;
 	}
